+++ nodes/__init__.py
@@ -1,7 +1,10 @@
 from .file_name_formatter import DossFileNameFormatter, format_doss_filename
+from .image_comparer import DossImageComparer, choose_comparison_images
 
 
 __all__ = [
     "DossFileNameFormatter",
+    "DossImageComparer",
+    "choose_comparison_images",
     "format_doss_filename",
 ]